Add content and styling
This adds content, at least a bit, and defines some okayish stling for
my website.  I am using my own site generator called ssg to render the website.
The rendered website will be published using GitHub pages.  This is the
easiest way to serve a static website and get HTTPS for free.

I have at least three older drafts for this website and tried a out a
bunch of things, e.g. one variant was a custom webserver with SQLite
storage, on-the-fly markdown rendering and multi-user account support.
I decided against this approach because it all causes a bit of
maintenance trouble I am not willing to spend my time with.  In the end
I just hope to write some articles, hopefully once a month at least.

All assets like CSS and fonts are bundled and nothing is included from
external service like Google Fonts.  This is for privacy and reliability
reasons.

The dark mode theme does not work for Chromium based browsers under
Linux because they do not support the 'prefers-color-scheme' media
query.