Add an article about learning how to touch type
I am still in the learning process and hope to be fluent in touch typing
at the end of November.  After a week of practicing about an hour per
day I am comfortable enough to use touch typing for my day job without
fearing that it slows me down to much.  Still, I am not fast and
accurate enough, so there is still some way to go.