Self host highlightjs
I was wondering why the website was loading so slowly and checked the
network tab in the browser devtools to see what is causing the delay.
It showed that loading from unpkg.com was very slow, sometimes multiple
seconds to iniate a request.  The only dependency that is loaded from
unpkg.com is highlightjs, which can be easily self-hosted by picking the
required files from their release build repository [1].

https://github.com/highlightjs/cdn-release